--- app/src/main/java/com/project/jangburich/ui/store/StoreFragment.kt
@@ -1,31 +1,0 @@
-package com.project.jangburich.ui.store
-
-import android.os.Bundle
-import android.util.Log
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.project.jangburich.databinding.FragmentStoreMapBinding
-import com.project.jangburich.ui.MainActivity
-
-class StoreMapFragment : Fragment() {
-
-    lateinit var binding: FragmentStoreMapBinding
-    lateinit var mainActivity: MainActivity
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        binding = FragmentStoreMapBinding.inflate(layoutInflater)
-        mainActivity = activity as MainActivity
-
-        mainActivity.hideBottomNavigation(false)
-
-
-        return binding.root
-    }
-
-
-}
